requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg[binary]>=3.2.2
psycopg2-binary==2.9.9
pandas>=2.2.0
openpyxl==3.1.2
python-dateutil==2.8.2
anthropic>=0.40.0
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.1
python-dotenv==1.0.0